chore: update project dependencies
a. Upgraded TypeScript from `^5` to `5.6.3` and ESLint from `^8` to `9.13.0`, along with related packages like `@types/node`, `@types/react`, and `@types/react-dom`.
b. This was done to address potential security vulnerabilities and ensure compatibility with the latest features and bug fixes, as well as to maintain up-to-date type definitions and avoid conflicts.
c. Updated the `pnpm-lock.yaml` file to reflect the new dependency versions and ensure reproducible builds.